requirements.txt
# serpapi==0.1.5
google_search_results==2.4.0
python-dotenv
pandas
streamlit
twilio
plotly
supabase